ensure complete writes
In practice, most implementations of `write` will completely write small
bytes, but it's ultimately up to the underlying implementation. If we
want to be sure the entire input is written out, we should use
`write_all`.

If we did want to use `write` we'd need some kind of recursion/looping
to be sure all the output is written.