Re-export prost::Message for mvt
An mvt tile is a protobuf message, parsed with the prost library.
That library gives us a trait to be able to

```rust
geozero::mvt::Tile::decode(raw_data)
```

However, the trait must be in the scope.

Until now, the solution was to have prost as an extra dependency and do

```rust
use prost::Message
```

This is a problem when the versions are not equivalent in a SemVer
sense as it might break randomly.

This commits allows the program that uses `geozero::mvt` to do

```rust
use geozero::mvt::Message
```

and no longer have an explicit dependency to prost, and we no longer
have the risk of version mismatch.