Make other wkb and wkt variants generic (#188)
This is a follow up to https://github.com/georust/geozero/pull/171 to
make other WKB and WKT variants generic over any `AsRef<[u8]>` input

With this change, I think `WktString` and `WktStr` are redundant, but I
didn't want to make a breaking change. I _think_ this should be
non-breaking because `String` and `&str` can `AsRef` to `[u8]` right?