Keep clippy happy with different feature flags to disable test (#228)
Clippy CI explodes because we're using a non-existent feature. It fails
to compile without DATABASE_URL

```
error: unexpected `cfg` condition value: `dont-compile`
   --> geozero/tests/postgis.rs:258:11
    |
258 |     #[cfg(feature = "dont-compile")]
    |           ^^^^^^^^^^^^^^^^^^^^^^^^
```

But what is this test testing anyway? We already have tests inserting
geo-points into sqlx (see rust_geo_query), so my understanding is that
the thing this test covers is that the sqlx macro behaves the same way,
as the programmatic API, which I'd argue is sqlx's job, not ours.