Add WrappedProcessor to help reading MVTs (#222)
My ultimate aim is to read an mbtiles file with a single zoom level and
extract features from it. It contains compressed MVT data. To that end,
I wrote
https://github.com/acteng/will-it-fit/blob/main/data_prep/fix_osmm/src/main.rs
using geozero and some other crates. It works like this:

1) Use the `mbtiles` crate to open the file
2) Calculate all tiles in the file
3) For each tile, read the raw data
4) Use `flate2` to gunzip it
5) Decode into the `geozero::mvt::Tile` proto
6) Use the existing mvt geozero reader to process all layers.
7) But because the geometry in each tile is scaled to fit that tile, I
need to transform the coordinates back into WGS84. IIUC the geozero
approach correctly, doing that translation as I go is more performant
than collecting into something like an FGB writer first and then trying
to go back and transform everything. So, I wrote `WrappedProcessor` to
delegate to another `FeatureProcessor` (an `FgbWriter` in this example),
but call a function on every coordinate first

This PR adds the `WrappedProcessor`, in case it might be helpful for
other use cases. But I'm not sure this approach is the nicest one --
maybe the
[mvt::process](https://github.com/georust/geozero/blob/c8a5f9103fc5ecc0ae9c7fcd2663b094e620da38/geozero/src/mvt/mvt_reader.rs#L17)
function should instead take optional info about the current tile (the
tile x and y, the zoom level, and the extent), plumb it through the
private methods in that file, and apply in
[process_coord](https://github.com/georust/geozero/blob/c8a5f9103fc5ecc0ae9c7fcd2663b094e620da38/geozero/src/mvt/mvt_reader.rs#L104)?

I'm not opinionated about the best way to read an mvt tile doing this
coordinate transformation, so happy to implement the other approach or
something else entirely. Thanks!

---------

Co-authored-by: Michael Kirk <michael.code@endoftheworl.de>