FIX: broken build, use local geozero dep.
Since geos 0.9 was yanked, geozero 0.14.0 no longer builds.

Previously, ./geozero-cli (and ./geozero-bench) depended not on the
geozero source code in the workspace, but on the published geozero@0.14.0

Because geozero-cli and geozero-bench could not build geozero@0.14.0, CI was failing.

At least for now, let's have geozero-cli and geozero-bench depend on the
local crate, because it fixed this build.

Personally, I think it would make sense to _keep_ crates in the
workspace using the local source code. But maybe others disagree.

In particular, I think I would be very surprised to have geozero-bench
actually bench the published geozero crate rather than whats in the
working directory.

And I think it's good to know when API changes to ./geozero require
changes to ./geozero-bench and ./geozero-cli

At the point you are making those changes in ./geozero is the best time
for you to understand the ramifications of those changes, addressing any
"downstream" breakages in the local workspace.