feat(ci): add codeql and cargo audit security tooling (#269)
- [x] I agree to follow the project's [code of
conduct](https://github.com/georust/.github/blob/main/CODE_OF_CONDUCT.md).
- [ ] I added an entry to the project's change log file if knowledge of
this change could be valuable to users.
  - Usually called `CHANGES.md` or `CHANGELOG.md`
  - Prefix changelog entries for breaking changes with "BREAKING: "
---

Currently, there is no security tooling in this project.
This PR adds the basic 
- codeql for checking for obvious issues
- cargo-adit for cecking that none of our dependencys have issues

I did not add it to the changelog as this is not something that users
should care about.
In the end, it is "just" more static analysis.